#2147 moved duplicated methods to CoreMainActivity
