+++ core/src/main/java/org/kiwix/kiwixmobile/core/main/CoreMainActivity.kt
@@ -19,8 +19,10 @@ package org.kiwix.kiwixmobile.core.main
 
 import android.content.Intent
 import android.os.Bundle
+import android.view.ActionMode
 import org.kiwix.kiwixmobile.core.R
 import org.kiwix.kiwixmobile.core.base.BaseActivity
+import org.kiwix.kiwixmobile.core.base.BaseFragmentActivityExtensions
 
 abstract class CoreMainActivity : BaseActivity(), WebViewProvider {
   override fun onCreate(savedInstanceState: Bundle?) {
@@ -41,6 +43,35 @@ abstract class CoreMainActivity : BaseActivity(), WebViewProvider {
     super.onRequestPermissionsResult(requestCode, permissions, grantResults)
     supportFragmentManager.fragments.forEach {
       it.onRequestPermissionsResult(requestCode, permissions, grantResults)
+    }
+  }
+
+  override fun onActionModeStarted(mode: ActionMode) {
+    super.onActionModeStarted(mode)
+    supportFragmentManager.fragments.filterIsInstance<BaseFragmentActivityExtensions>().forEach {
+      it.onActionModeStarted(mode, this)
+    }
+  }
+
+  override fun onActionModeFinished(mode: ActionMode) {
+    super.onActionModeFinished(mode)
+    supportFragmentManager.fragments.filterIsInstance<BaseFragmentActivityExtensions>().forEach {
+      it.onActionModeFinished(mode, this)
+    }
+  }
+
+  override fun onBackPressed() {
+    supportFragmentManager.fragments.filterIsInstance<BaseFragmentActivityExtensions>().forEach {
+      if (it.onBackPressed(this) == BaseFragmentActivityExtensions.Super.ShouldCall) {
+        super.onBackPressed()
+      }
+    }
+  }
+
+  override fun onNewIntent(intent: Intent) {
+    super.onNewIntent(intent)
+    supportFragmentManager.fragments.filterIsInstance<BaseFragmentActivityExtensions>().forEach {
+      it.onNewIntent(intent, this)
     }
   }
